requirements/base.requirements.txt
# app specific
Django
psycopg2  
Pillow  
django-heroku  
requests  
boto3
django-storages  
inflection
graphql-core 
graphene
graphene-django 
graphene-file-upload
django-graphql-jwt
django-filter
celery
django-celery-results
django-celery-beat
django-timezone-field
djangorestframework
jsonfield==2.1.1
./lib/django-mutant-0.3a5.tar.gz
django-polymodels>=1.5.1
rdflib
redis
django-modeladmin-reorder
django-celery-model
model-bakery
parameterized
django-cors-headers

flower
./lib/django_json2model-0.2.7-py3-none-any.whl
./lib/django_generic_serializer-0.0.2-py3-none-any.whl
gunicorn

# generic
invoke
python-dotenv



# dummy change 2
